refactor: update include paths for object association modules
